Time: 77 ms (87.5%), Space: 198.6 MB (72.02%) - LeetHub

+++ 3453-separate-squares-i/3453-separate-squares-i.cpp
@@ -12,8 +12,8 @@ public:
         }
 
         double ans = miny + (maxy-miny)/2.0;
-        long double ansTop = maxy;
-        long double ansBottom = miny;
+        double ansTop = maxy;
+        double ansBottom = miny;
 
         bool cut = false;
